Multiple host functionality for offline compression
The ability to change the host from which to serve a static file is required,
instead of injecting the contents of the manifest file into the template,
re-render it at request time based using information from the manifest file.

Templates can be added to your project that override those supplied here, so you
can easily add your own method for finding the correct static host to suit
your needs!